requirements.txt
streamlit
pandas
matplotlib
seaborn
plotly
plotly.express
openpyxl
msoffcrypto-tool
